Add RESTful controllers, models and route them
We prefer use RESTful to pass data to page. In fact, we want to use AJAX
to load data by REST, that why we want to create RESTful controllers but
not normal one. The reason what we want use the way because we wanna
create a single-page application.

To correspond tables, we create 3 controllers and models:

* users / UserController / User
* boards / BoardController / Board
* apply_record / ApplyRecordController / ApplyRecord

abcdefghijklmnopqrstuvwxyzabcdefghijklmnopqrstuvw1234567890123456789012345